Removed mixin and replaced with fragment extension.
